pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-lab"
version = "0.1.0"
description = "AI-Powered Quantitative Trading Strategy Development Tool"
authors = [{name = "Quant Lab Team", email = "team@quantlab.ai"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "shap>=0.42.0",
    "optuna>=3.2.0",
    "mlflow>=2.6.0",
    "pyarrow>=12.0.0",
    "pydantic>=2.0.0",
    "hydra-core>=1.3.0",
            "pyyaml>=6.0",
        "python-dotenv>=1.0.0",
        "matplotlib>=3.7.0",
        "plotly>=5.15.0",
        "yfinance>=0.2.18",
        "alpha-vantage>=2.3.1",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.3.0",
    "mypy>=1.5.0",
]

[project.scripts]
quant-lab = "quant_lab.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
